mostly done with static and dynamic normalization visualization
-Updated the way public install cost bars update when switched between
dynamic and static normalization

-when using static normalization, public install cost bars are
visualized in terms of the max budget (if install cost is within the
budget it fill in green with some gray if a bit of the budget is left)
while filling in red when it is over budget.

-When using Dynamic normalization, the public install cost bars grow
dynamically to the size of the highest installation cost found so far.
Thus, a red bar will never appear. Also, the trial with the highest
installation cost will not have a remainder left of the budget (no gray
bar) and the rest of the trials will have some remainder (some gray
bars)