refactored methods responsible for drawing and updating
installment cost displays. methods now can handle static and dynamic
normalization writing and updating with little no difference in
arguments.

biggest difference is that it takes a highestCost argument  (intended
to be the highest cost found during dynamic normalization) which serves
a flag to determine whether the public display is going to be done
dynamically or statically.  that parameter will be set to -1 if the
update is a static one.