Major changes in public cost display visualization methods
simplified it a bit and stopped using the normalized values to draw the
visualizations. moved on to computing the actual width that will be
taken up by a particular value on the public score bar and passed them
as parameters. If the green bar’s greater than the max budget’s bar
width, the difference will be calculated and drawn over as red.