found interesting bug and then fixed it
Overtime that you leave the outcomesalienceview and go back to it, the
title view gets reset back to its original position (looking at the
first concern ranking) even when the contents of the data window (where
all the trial information is) is offset somewhere else.

this was never an issue since the contents of the view were removed and
redrawn overtime. but now that they aren’t, it was a visible bug. this
was fixed, however.